requirements.txt
langchain
chromadb
pdfplumber
python-dotenv 